+++ src/Messages/Read/MessageReadKind.ts
@@ -10,5 +10,7 @@ export enum MessageReadKind {
     activeDotNetSolutionFile,
     nugetPackageReferencesInProject,
     undefinedSolution,
-    requestForRefresh
+    requestForRefresh,
+    filtersInVCXProject,
+    vcxFilterMatches
 }