+++ svelteApps/solutionExplorerWebview/src/Components/TreeViews/CPlusPlus/VCXProjectTreeView.svelte
@@ -2,7 +2,12 @@
 	import type { IdeFile } from "../../../../../../out/FileSystem/Files/IdeFile";
 	import type { VCXProjectFile } from "../../../../../../out/FileSystem/Files/CPlusPlus/VCXProjectFile";
 	import { MessageReadFileIntoEditor } from "../../../../../../out/Messages/Read/MessageReadFileIntoEditor";
+	import { MessageReadFiltersInVCXProject } from "../../../../../../out/Messages/Read/MessageReadFiltersInVCXProject";
 	import TreeViewBase from "../../TreeViewBase.svelte";
+	import type { MessageReadRequestForRefresh } from "../../../../../../out/Messages/Read/MessageReadRequestForRefresh";
+	import { onMount } from "svelte";
+	import { MessageCategory } from "../../../../../../out/Messages/MessageCategory";
+	import { MessageReadKind } from "../../../../../../out/Messages/Read/MessageReadKind";
 	
     export let vcxProjectFile: VCXProjectFile;
 
@@ -20,7 +25,20 @@
     }
 
 	function getChildFiles(): IdeFile[] {
-		return vcxProjectFile.constantChildFiles;
+		children = vcxProjectFile.virtualChildFiles;
+		
+		if (!children) {
+			let messageReadFiltersInVCXProject =
+				new MessageReadFiltersInVCXProject(vcxProjectFile);
+			
+			tsVscode.postMessage({
+				type: undefined,
+				value: messageReadFiltersInVCXProject,
+			});
+			return [];
+		}
+
+		return children;
 	}
 
 	function hasDifferentParentContainer(childIdeFile: IdeFile): boolean {
@@ -32,6 +50,48 @@
 
 		return false;
 	}
+	
+	onMount(async () => {
+		window.addEventListener("message", async (event) => {
+			const message = event.data;
+			switch (message.messageCategory) {
+				case MessageCategory.read:
+					switch (message.messageReadKind) {
+						case MessageReadKind.filtersInVCXProject:
+							let messageReadFiltersInVCXProject =
+								message as MessageReadFiltersInVCXProject;
+							if (
+								vcxProjectFile.nonce ===
+									messageReadFiltersInVCXProject.vcxProjectFile.nonce
+							) {
+								vcxProjectFile =
+									messageReadFiltersInVCXProject.vcxProjectFile;
+
+								children = vcxProjectFile.virtualChildFiles;
+							}
+							break;
+					}
+					case MessageReadKind.requestForRefresh:
+							let messageReadRequestForRefresh =
+								message as MessageReadRequestForRefresh;
+								
+							if (vcxProjectFile.nonce ===
+									messageReadRequestForRefresh.ideFileNonce) {
+								
+									let messageReadFiltersInVCXProject =
+										new MessageReadFiltersInVCXProject(
+											vcxProjectFile
+										);
+
+									tsVscode.postMessage({
+										type: undefined,
+										value: messageReadFiltersInVCXProject,
+									});
+							}
+							break;
+			}
+		});
+	});
 </script>
 
 <TreeViewBase ideFile="{vcxProjectFile}"